requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
wandb>=0.16.0
einops>=0.7.0
mamba-ssm>=1.0.1
causal-conv1d>=1.0.0
datasets>=2.14.0
transformers>=4.36.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
pandas>=2.1.0
hydra-core>=1.3.0
omegaconf>=2.3.0
tensorboard>=2.14.0
pytorch-lightning>=2.1.0